stage0: move OSTree logic into dedicated module
